move discord events to mod file
